requirements.txt
fire==0.5.0
markdown-it-py==2.2.0
mdurl==0.1.2
Pygments==2.14.0
python-frontmatter==1.0.0
PyYAML==6.0
rich==13.3.1
six==1.16.0
termcolor==2.2.0
